pyproject.toml
[project]
name = "kokoromemo"
version = "0.1.0"
description = "Local long-term memory proxy for AI role-playing applications"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "lancedb>=0.17",
    "pyarrow>=17",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
